admin auth done

+++ src/manage/manage.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { firestore } from "../firebase";
-import { collection, getDocs, doc, deleteDoc } from "firebase/firestore";
+import { collection, getDocs, doc, deleteDoc ,getDoc } from "firebase/firestore";
+import { getAuth, deleteUser } from "firebase/auth";
 import {
   Typography,
   Box,
@@ -76,24 +77,40 @@ function Manage() {
     navigate("/manage/add");
   };
 
-const handleDeleteUser = async (uid) => {
-  if (uid === "Syo5nn3QB0RxgRTwSN34ImkMMKp1") {
-    alert("ไม่สามารถลบข้อมูลนี้ได้");
-    return;
-  }
-
-  const confirmed = window.confirm("โปรดยืนยันการลบข้อมูล");
-  if (confirmed) {
-    try {
-      await deleteDoc(doc(firestore, "admins", uid));
-      const updatedUserData = userData.filter((user) => user.uid !== uid);
-      setUserData(updatedUserData);
-    } catch (error) {
-      console.error("Error deleting user: ", error);
+  const handleDeleteUser = async (uid) => {
+    if (uid === "Syo5nn3QB0RxgRTwSN34ImkMMKp1") {
+      alert("ไม่สามารถลบข้อมูลนี้ได้");
+      return;
     }
-  }
-};
 
+    const confirmed = window.confirm("โปรดยืนยันการลบข้อมูล");
+    if (confirmed) {
+      try {
+        const currentEmail = localStorage.getItem("email");
+    
+        const adminDocRef = doc(firestore, "admins", uid);
+        const adminDocSnapshot = await getDoc(adminDocRef);
+        if (adminDocSnapshot.exists()) {
+          const adminData = adminDocSnapshot.data();
+          if (currentEmail === adminData.email) {
+            await deleteDoc(adminDocRef);
+            localStorage.removeItem('authToken');
+            localStorage.removeItem('email');
+            alert("ลบข้อมูลสำเร็จ");
+            navigate("/login");
+          } else {
+            await deleteDoc(doc(firestore, "admins", uid));
+            alert("ลบข้อมูลสำเร็จ");
+            window.location.reload();
+          }
+        } else {
+          alert("ไม่พบข้อมูลผู้ใช้");
+        }
+      } catch (error) {
+        console.error("Error deleting user: ", error);
+      }
+    }
+  };
 
   return (
     <div>
@@ -101,78 +118,80 @@ const handleDeleteUser = async (uid) => {
       <Box sx={{ marginLeft: 25 }}>
         <Card variant="outlined">
           <CardContent>
-          <form>
-            <Box sx={{ overflow: { xs: "auto", sm: "unset" } }}>
-              <Table aria-label="simple table" className={classes.table}>
-                <TableHead>
-                  <TableCell style={{ backgroundColor: "transparent" }}>
-                    <Typography
-                      variant="subtitle1"
-                      style={{
-                        fontSize: "13px",
-                        fontWeight: "bold",
-                        fontFamily: "Arial, sans-serif",
-                        color: "#4a5568",
-                      }}
-                    >
-                      ID
-                    </Typography>
-                  </TableCell>
+            <form>
+              <Box sx={{ overflow: { xs: "auto", sm: "unset" } }}>
+                <Table aria-label="simple table" className={classes.table}>
+                  <TableHead>
+                    <TableCell style={{ backgroundColor: "transparent" }}>
+                      <Typography
+                        variant="subtitle1"
+                        style={{
+                          fontSize: "13px",
+                          fontWeight: "bold",
+                          fontFamily: "Arial, sans-serif",
+                          color: "#4a5568",
+                        }}
+                      >
+                        ID
+                      </Typography>
+                    </TableCell>
 
-                  <TableCell style={{ backgroundColor: "transparent" }}>
-                    <Typography
-                      variant="subtitle1"
-                      style={{
-                        fontSize: "12px",
-                        fontWeight: "bold",
-                        fontFamily: "Arial, sans-serif",
-                        color: "#4a5568",
-                      }}
-                    >
-                      EMAIL
-                    </Typography>
-                  </TableCell>
+                    <TableCell style={{ backgroundColor: "transparent" }}>
+                      <Typography
+                        variant="subtitle1"
+                        style={{
+                          fontSize: "12px",
+                          fontWeight: "bold",
+                          fontFamily: "Arial, sans-serif",
+                          color: "#4a5568",
+                        }}
+                      >
+                        EMAIL
+                      </Typography>
+                    </TableCell>
 
-                  <TableCell style={{ backgroundColor: "transparent" }}>
-                    <Typography
-                      variant="subtitle1"
-                      style={{
-                        fontSize: "12px",
-                        fontWeight: "bold",
-                        fontFamily: "Arial, sans-serif",
-                        color: "#4a5568",
-                      }}
-                    >
-                      ACTION
-                    </Typography>
-                  </TableCell>
-                </TableHead>
-                <TableBody>
-                  {userData.map((user, index) => (
-                    <TableRow key={index}>
-                      <TableCell>{user.uid || "N/A"}</TableCell>
-                      <TableCell>{user.email || "N/A"}</TableCell>
-                      <TableCell>
-                        <IconButton onClick={() => handleDeleteUser(user.uid)}>
-                          <DeleteIcon />
-                        </IconButton>
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </Box>{" "}
-            <Box textAlign="left">
-              <Button
-               style={{marginTop:"8px"}}
-                color="primary"
-                variant="contained"
-                type="button"
-                onClick={handleAddUser}
-              >
-                เพิ่ม
-              </Button>
-            </Box>
+                    <TableCell style={{ backgroundColor: "transparent" }}>
+                      <Typography
+                        variant="subtitle1"
+                        style={{
+                          fontSize: "12px",
+                          fontWeight: "bold",
+                          fontFamily: "Arial, sans-serif",
+                          color: "#4a5568",
+                        }}
+                      >
+                        ACTION
+                      </Typography>
+                    </TableCell>
+                  </TableHead>
+                  <TableBody>
+                    {userData.map((user, index) => (
+                      <TableRow key={index}>
+                        <TableCell>{user.uid || "N/A"}</TableCell>
+                        <TableCell>{user.email || "N/A"}</TableCell>
+                        <TableCell>
+                          <IconButton
+                            onClick={() => handleDeleteUser(user.uid)}
+                          >
+                            <DeleteIcon />
+                          </IconButton>
+                        </TableCell>
+                      </TableRow>
+                    ))}
+                  </TableBody>
+                </Table>
+              </Box>{" "}
+              <Box textAlign="left">
+                <Button
+                  style={{ marginTop: "8px" }}
+                  color="primary"
+                  variant="contained"
+                  type="button"
+                  onClick={handleAddUser}
+                >
+                  เพิ่ม
+                </Button>
+              </Box>
             </form>
           </CardContent>
         </Card>
